refactored user service: add exception hadler and changed verifications

+++ src/main/java/com/example/testproject/controller/UserController.java
@@ -1,21 +1,21 @@
 package com.example.testproject.controller;
 
 import com.example.testproject.entity.UserEntity;
-import com.example.testproject.exception.CustomBadRequestException;
-import com.example.testproject.exception.CustomNotFoundException;
-import com.example.testproject.model.common.ApiError;
+import com.example.testproject.exception.SystemException;
+import com.example.testproject.model.common.ApiResponse;
 import com.example.testproject.model.common.ApiResponseError;
 import com.example.testproject.model.user.User;
-import com.example.testproject.model.user.UserApiResponse;
+import com.example.testproject.model.user.UsersList;
 import com.example.testproject.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.persistence.NonUniqueResultException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api/v1/users")
@@ -25,53 +25,62 @@ public class UserController {
     private UserService userService;
 
     @PostMapping("/user")
-    public ResponseEntity registerUsers(@RequestBody UserEntity user) {
-        List<ApiError> errors = new ArrayList<>();
-        try {
-            userService.createUser(user);
-            return ResponseEntity.ok(UserApiResponse.buildApiResponseObject(User.toModel(user)));
-        } catch (CustomBadRequestException e) {
-            errors.add(ApiError.buildApiError(HttpStatus.BAD_REQUEST.getReasonPhrase(), e.getMessage()));
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ApiResponseError.buildApiResponseError(errors));
+    public ResponseEntity<?> registerUsers(@RequestBody UserEntity user) {
+        List<ApiResponseError> errors = new ArrayList<>();
+        if (user.getEmail() == null) {
+            errors.add(ApiResponseError.badRequest("Field is required: email"));
+            return ResponseEntity.badRequest().body(ApiResponse.apiErrors(errors));
         }
+        if (user.getPassword() == null) {
+            errors.add(ApiResponseError.badRequest("Field is required: password"));
+            return ResponseEntity.badRequest().body(ApiResponse.apiErrors(errors));
+        }
+        if (userService.findUser(user.getEmail()) != null) {
+            String message = String.format("User already exists, email: %s", user.getEmail());
+            throw new NonUniqueResultException(message);
+        }
+        User newUser = User.toUserModel(userService.createUser(user));
+        return ResponseEntity.ok(ApiResponse.okContent(newUser));
     }
 
     @GetMapping("/user/{id}")
-    public ResponseEntity getUserInfo(@PathVariable Long id) {
-        List<ApiError> errors = new ArrayList<>();
-        try {
-            UserEntity user = userService.getUser(id);
-            return ResponseEntity.ok(User.toModel(user));
-        } catch (CustomNotFoundException e) {
-            errors.add(ApiError.buildApiError(HttpStatus.NOT_FOUND.getReasonPhrase(), e.getMessage()));
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponseError.buildApiResponseError(errors));
+    public ResponseEntity<?> getUserInfo(@PathVariable Long id) {
+        List<ApiResponseError> errors = new ArrayList<>();
+        if (userService.findUser(id) == null) {
+            errors.add(ApiResponseError.notFound(String.format("User with id %d not found", id)));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.apiErrors(errors));
         }
+        if (Long.parseLong(id.toString()) == 0) {
+            String message = String.format("Invalid input: %s", id);
+            throw new NonUniqueResultException(message);
+        }
+        return ResponseEntity.ok(ApiResponse.okContent(User.toUserModel(userService.findUser(id))));
     }
 
     @GetMapping("/list")
-    public ResponseEntity getAllUsers() {
-        List<ApiError> errors = new ArrayList<>();
+    public ResponseEntity<?> getUsersList() {
+        List<ApiResponseError> errors = new ArrayList<>();
         try {
-            List<User> users = userService.getUsers();
-            return ResponseEntity.ok(UserApiResponse.buildApiResponseList(users));
-        } catch (Exception e) {
-            errors.add(ApiError.buildApiError(HttpStatus.BAD_REQUEST.getReasonPhrase(),
-                    userService.getMessageSource().getMessage("system.error", new Object[0],
-                            new Locale("en"))));
-            return ResponseEntity.badRequest().body(ApiResponseError.buildApiResponseError(errors));
+            List<User> users = userService.findAllUsers().stream().map(User::toUserModel).collect(Collectors.toList());
+            return ResponseEntity.ok(ApiResponse.okContent(UsersList.builder()
+                    .users(users)
+                    .count(users.size())
+                    .build()));
+        } catch (SystemException e) {
+            errors.add(ApiResponseError.badRequest(e.getMessage()));
+            return ResponseEntity.badRequest().body(ApiResponse.apiErrors(errors));
         }
     }
 
     @DeleteMapping("/user/{id}")
-    public ResponseEntity deleteUser(@PathVariable Long id) {
-        List<ApiError> errors = new ArrayList<>();
-        try {
-            userService.removeUser(id);
-            return ResponseEntity.ok(userService.getMessageSource().getMessage("operation.success", new Object[0],
-                    new Locale("en")));
-        } catch (CustomNotFoundException e) {
-            errors.add(ApiError.buildApiError(HttpStatus.NOT_FOUND.getReasonPhrase(), e.getMessage()));
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponseError.buildApiResponseError(errors));
+    public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+        List<ApiResponseError> errors = new ArrayList<>();
+        if (userService.findUser(id) == null) {
+            errors.add(ApiResponseError.notFound(String.format("User with id %d not found", id)));
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.apiErrors(errors));
         }
+        userService.removeUser(id);
+        return ResponseEntity.ok(String.format("User with id: %d was removed successfully", id));
+
     }
 }
